Changes from static context to instance context and adds a new class, MathHelper.java.
The application was using static methods and fields. This patch changes that to use of instance methods and fields to add flexibility to the code. It also adds a new class, src/MathHelper.java, with the aim to group related components of the application together to make the code shorter and increase readability. Also to shorten the code even further, it removes the use of StringBuilder class and the many calls that were being daisy-chained to its append() method when printing results of a calculation to the console, in class src/Data.java, and replaced it with a single-line literal string.