Fixes a bug
When the user performed first calculation and went ahead to do a second calculation using result of first calculation but a different operator to the operator used in the first calculation, the application was ignoring the user's choice of a different operator and was reusing the initial operator. This change set fixes that by setting the operator to be used in the subsequent calculation to user decision that s/he types in.